Edge rim outlining implemented; Timer code refined; Timer rawdata produced
