+++ src/renderer.ts
@@ -25,6 +25,9 @@ export async function initWebGPU() {
     canvas.width = canvas.clientWidth * devicePixelRatio;
     canvas.height = canvas.clientHeight * devicePixelRatio;
 
+    console.log(canvas.clientWidth);
+    console.log(canvas.clientHeight);
+
     aspectRatio = canvas.width / canvas.height;
 
     if (!navigator.gpu)
@@ -126,8 +129,15 @@ export abstract class Renderer {
 
     protected gpuTime: number = 0;
     protected gpuTimes: Array<number>;
+    protected gpuTimesPre: Array<number>;
+    protected gpuTimesPost: Array<number>;
+    protected gpuTimesCompute: Array<number>;
     protected gpuTimesIndex: number = 0;
     protected gpuTimesSize: number = 100;
+
+    // IMPORTANT: Edit these flags to log times on browser console
+    private logTime = false;
+    protected logSeparateTimes = true;
 
     constructor(stage: Stage) {
         this.scene = stage.scene;
@@ -139,10 +149,10 @@ export abstract class Renderer {
 
         this.querySet = device.createQuerySet({
             type: 'timestamp',
-            count: 4,
+            count: 6,
         });
         this.resolveBuffer = device.createBuffer({
-            size: this.querySet.count * 8,
+            size: this.querySet.count * 12,
             usage: GPUBufferUsage.QUERY_RESOLVE | GPUBufferUsage.COPY_SRC,
         });
         this.resultBuffer = device.createBuffer({
@@ -151,6 +161,11 @@ export abstract class Renderer {
         });
 
         this.gpuTimes = new Array<number>(this.gpuTimesSize);
+        this.gpuTimesPre = new Array<number>(this.gpuTimesSize);
+        this.gpuTimesPost = new Array<number>(this.gpuTimesSize);
+        this.gpuTimesCompute = new Array<number>(this.gpuTimesSize);
+
+        canTimestamp = canTimestamp && this.logTime;
     }
 
     stop(): void {
