+++ src/renderers/forward_plus.ts
@@ -172,6 +172,7 @@ export class ForwardPlusRenderer extends renderer.Renderer {
         }
 
         if (this.gpuTimesIndex == this.gpuTimesSize) {
+            console.log("Overall time");
             console.log(this.gpuTimes);
             this.gpuTimesIndex++;
         }
